Hosting: Add built files to repo.
Since we want to use static-site hosting, we need to have the built
website checked into the code base.

This will be a bit annoying since every change will mess with index.js,
and will mess with git conflicts, but it's necessary to not get charged
to build the website, and have it hosted for free.

When git conflicts happen, resolve the usual conflicts, and then delete
the dist folder, run pnpm run build, and then commit those new files.